requirements.txt
# Core dependencies
# fastapi
# uvicorn
# gunicorn
#
# pydantic
# pydantic-settings
# python-multipart
#
# # LLM and LangChain related
# langchain
# langchain-openai
# langgraph
#
# mem0ai
#
# # HTTP client
# httpx
#
# # Testing
# pytest
#
# # Web UI
# flask
# requests

aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
backoff==2.2.1
blinker==1.9.0
brotli==1.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
daal==2025.5.0
distro==1.9.0
fastapi==0.121.2
ffmpy==1.0.0
filelock==3.20.0
Flask==3.1.2
fsspec==2025.10.0
gradio==5.49.1
gradio_client==1.13.3
greenlet==3.2.4
groovy==0.1.2
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.1.4
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.7
langchain-core==1.0.5
langchain-openai==1.0.3
langchain-text-splitters==0.3.11
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.43
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mem0ai==1.0.1
numpy==2.3.5
openai==2.8.0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pip-autoremove==0.10.0
pluggy==1.6.0
portalocker==3.2.0
posthog==7.0.1
protobuf==5.29.5
pydantic==2.11.10
pydantic-settings==2.12.0
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pytest==9.0.1
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.3
qdrant-client==1.15.1
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
ruff==0.14.5
safehttpx==0.1.7
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
tenacity==9.1.2
tiktoken==0.12.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
xxhash==3.6.0
zstandard==0.25.0
